requirements.txt
streamlit
pandas
plotly-express
cryptography
python-dotenv
firebase-admin
sqlalchemy